Split WordDetail component
